requirements.txt
numpy==1.21.6
python-dotenv==0.21.0
python_telegram_bot==13.14
telegram==0.0.1
torch==1.13.0
